requirements.txt
pandas>=1.5.0
numpy>=1.23.0
requests>=2.31.0
PyYAML>=6.0
tensorflow>=2.14.0
xgboost>=1.7.6
arch>=5.3.1
ta>=0.10.2
pytest>=7.4.0
